backend/requirements.txt
fastapi==0.115.6
uvicorn[standard]==0.34.0
pydantic==2.10.4
pydantic-settings==2.7.1
SQLAlchemy==2.0.36
asyncpg==0.30.0
alembic==1.14.0
python-dotenv==1.0.0
greenlet>=3.0.0
